+++ R/bind_tibbles.R
@@ -32,9 +32,7 @@
 #' }
 #' @export
 
-bind_tibbles <- function(supertbl,
-                         environment = global_env(),
-                         tbls = NULL) {
+bind_tibbles <- function(supertbl, environment = global_env(), tbls = NULL) {
   check_arg_is_supertbl(supertbl, req_cols = "redcap_data")
   check_arg_is_env(environment)
   check_arg_is_character(tbls, null.ok = TRUE, any.missing = FALSE, min.len = 1)
@@ -60,5 +58,5 @@ bind_tibbles <- function(supertbl,
       value = .y
     )
   )
-  return(invisible(NULL))
+  invisible(NULL)
 }